ci: fix vega/cml pipe communication
Before this commit sync issues between vega and cml
commands could happen.

```
vega input args | cml-publish args >> file.md
```

If anything goes wrong with vl2png or the stream of
the png is not finished before  cml-publish call this
will trigger a failure in the cml-publish. Thus, causing
failure in the vl2png process because there is no place
to stream the png file